Support more than just hex for color type.
It's a little messy, so I welcome anyone to improve it, but this adjustment allows common names, hex values (with and without #), RANDOM, RGB array, and numbers to be used as values for the color ArgumentType.

Note that when using an RGB array, either the array must not have spaces in it, or `quoted: true` must be applied to the argument.

This method takes advantage of the resolveColor() method from discord.js. This theoretically means that anything that can be considered ColorResolvable should work.